Remove Request suffix from  models
These models will be used for response models going forward as well, so it
makes sense to drop the `Request` suffix from their name.
