+++ GetIntoTeachingApi/Models/MailingListAddMember.cs
@@ -3,19 +3,24 @@ using System.Text.Json.Serialization;
 
 namespace GetIntoTeachingApi.Models
 {
-    public class TeachingEventAddAttendeeRequest
+    public class MailingListAddMember
     {
         public Guid? CandidateId { get; set; }
-        public Guid EventId { get; set; }
+        public Guid? QualificationId { get; set; }
+        public Guid PreferredTeachingSubjectId { get; set; }
         public Guid AcceptedPolicyId { get; set; }
+
+        public int DescribeYourselfOptionId { get; set; }
+        public int ConsiderationJourneyStageId { get; set; }
+        public int UkDegreeGradeId { get; set; }
 
         public string Email { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string AddressPostcode { get; set; }
         public string Telephone { get; set; }
+        public string CallbackInformation { get; set; }
         public bool SubscribeToEvents { get; set; }
-        public bool SubscribeToMailingList { get; set; }
 
         [JsonIgnore]
         public Candidate Candidate => CreateCandidate();
@@ -25,36 +30,32 @@ namespace GetIntoTeachingApi.Models
             var candidate = new Candidate()
             {
                 Id = CandidateId,
+                DescribeYourselfOptionId = DescribeYourselfOptionId,
+                ConsiderationJourneyStageId = ConsiderationJourneyStageId,
+                PreferredTeachingSubjectId = PreferredTeachingSubjectId,
                 Email = Email,
                 FirstName = FirstName,
                 LastName = LastName,
                 AddressPostcode = AddressPostcode,
                 Telephone = Telephone,
+                CallbackInformation = CallbackInformation,
                 PrivacyPolicy = new CandidatePrivacyPolicy() { AcceptedPolicyId = AcceptedPolicyId },
-                ChannelId = CandidateId == null ? (int?)Candidate.Channel.Event : null,
+                ChannelId = CandidateId == null ? (int?)Candidate.Channel.MailingList : null,
                 EligibilityRulesPassed = "false",
                 OptOutOfSms = false,
-                DoNotBulkEmail = true,
+                DoNotBulkEmail = false,
                 DoNotEmail = false,
                 DoNotBulkPostalMail = true,
                 DoNotPostalMail = true,
-                DoNotSendMm = true,
+                DoNotSendMm = false,
             };
 
-            candidate.TeachingEventRegistrations.Add(new TeachingEventRegistration()
-            {
-                EventId = EventId,
-                ChannelId = (int)TeachingEventRegistration.Channel.Event,
-            });
+            candidate.Qualifications.Add(new CandidateQualification() { Id = QualificationId, UkDegreeGradeId = UkDegreeGradeId });
+            candidate.Subscriptions.Add(new Subscription() { TypeId = (int)Subscription.ServiceType.MailingList });
 
             if (SubscribeToEvents)
             {
                 candidate.Subscriptions.Add(new Subscription() { TypeId = (int)Subscription.ServiceType.Event });
-            }
-
-            if (SubscribeToMailingList)
-            {
-                candidate.Subscriptions.Add(new Subscription() { TypeId = (int)Subscription.ServiceType.MailingList });
             }
 
             return candidate;
