+++ GetIntoTeachingApi/Models/Validators/TeachingEventAddAttendeeValidator.cs
@@ -3,9 +3,9 @@ using GetIntoTeachingApi.Services;
 
 namespace GetIntoTeachingApi.Models.Validators
 {
-    public class TeacherTrainingAdviserSignUpRequestValidator : AbstractValidator<TeacherTrainingAdviserSignUpRequest>
+    public class TeachingEventAddAttendeeValidator : AbstractValidator<TeachingEventAddAttendee>
     {
-        public TeacherTrainingAdviserSignUpRequestValidator(IStore store)
+        public TeachingEventAddAttendeeValidator(IStore store)
         {
             RuleFor(request => request.Candidate).SetValidator(new CandidateValidator(store));
         }
