+++ GetIntoTeachingApiTests/Models/Validators/TeachingEventAddAttendeeValidatorTests.cs
@@ -1,47 +1,41 @@
-﻿using System;
-using System.Linq;
-using FluentAssertions;
+﻿using FluentAssertions;
 using FluentValidation.TestHelper;
 using GetIntoTeachingApi.Models;
 using GetIntoTeachingApi.Models.Validators;
 using GetIntoTeachingApi.Services;
 using Moq;
+using System;
+using System.Linq;
 using Xunit;
 
 namespace GetIntoTeachingApiTests.Models.Validators
 {
-    public class MailingListAddMemberRequestValidatorTests
+    public class TeachingEventAddAttendeeValidatorTests
     {
-        private readonly MailingListAddMemberRequestValidator _validator;
+        private readonly TeachingEventAddAttendeeValidator _validator;
         private readonly Mock<IStore> _mockStore;
 
-        public MailingListAddMemberRequestValidatorTests()
+        public TeachingEventAddAttendeeValidatorTests()
         {
             _mockStore = new Mock<IStore>();
-            _validator = new MailingListAddMemberRequestValidator(_mockStore.Object);
+            _validator = new TeachingEventAddAttendeeValidator(_mockStore.Object);
         }
 
         [Fact]
         public void Validate_WhenValid_HasNoErrors()
         {
-            var mockPickListItem = new TypeEntity { Id = "123" };
-            var mockEntityReference = new TypeEntity { Id = Guid.NewGuid().ToString() };
             var mockPrivacyPolicy = new PrivacyPolicy { Id = Guid.NewGuid() };
 
-            var request = new MailingListAddMemberRequest()
+            var request = new TeachingEventAddAttendee()
             {
                 CandidateId = null,
-                PreferredTeachingSubjectId = Guid.Parse(mockEntityReference.Id),
+                EventId = Guid.NewGuid(),
                 AcceptedPolicyId = (Guid)mockPrivacyPolicy.Id,
-                DescribeYourselfOptionId = int.Parse(mockPickListItem.Id),
-                ConsiderationJourneyStageId = int.Parse(mockPickListItem.Id),
-                UkDegreeGradeId = int.Parse(mockPickListItem.Id),
                 Email = "email@address.com",
                 FirstName = "John",
                 LastName = "Doe",
                 Telephone = "1234567",
                 AddressPostcode = "KY11 9YU",
-                CallbackInformation = "Test information",
             };
 
             var result = _validator.TestValidate(request);
@@ -55,20 +49,14 @@ namespace GetIntoTeachingApiTests.Models.Validators
         [Fact]
         public void Validate_CandidateIsInvalid_HasError()
         {
-            var request = new MailingListAddMemberRequest
+            var request = new TeachingEventAddAttendee
             {
-                PreferredTeachingSubjectId = Guid.NewGuid(),
+                FirstName = null,
             };
 
             var result = _validator.TestValidate(request);
 
-            result.ShouldHaveValidationErrorFor("Candidate.PreferredTeachingSubjectId");
-        }
-
-        [Fact]
-        public void Validate_AddressPostcodeIsNull_HasError()
-        {
-            _validator.ShouldHaveValidationErrorFor(request => request.AddressPostcode, null as string);
+            result.ShouldHaveValidationErrorFor("Candidate.FirstName");
         }
     }
 }
